refactor(kwin-effect): extract state helpers to autotilehandler/state.cpp
Phase 3 of the file split. Adds a third partition under
autotilehandler/, alongside the existing signals.cpp and tiling.cpp.

state.cpp carries 13 state-mutation and query helpers:
transferPreAutotileGeometry, unmaximizeMonocleWindow,
restoreAllMonocleMaximized, restoreAllBorderless,
updateHideTitleBarsSetting, updateShowBorderSetting,
setFocusFollowsMouse, saveAndRecordPreAutotileGeometry,
setWindowBorderless, isAutotileScreen, savePreAutotileForDesktopMove,
isEligibleForAutotileNotify, applyFloatCleanup.

autotilehandler.cpp drops from 981 to 655 lines.
